"master"

+++ resource/cog.py
@@ -1,5 +1,4 @@
 import discord
-from discord import app_commands
 from discord.ext import commands
 import datetime
 import time
@@ -20,20 +19,28 @@ class command(commands.Cog):
         self.bot=bot
         self.logch=None
         self.psql=psycopg2.connect(host=HOST,user=USERS,password=PASSWORD,database=DATABASE)
+        with self.psql.cursor() as cursor:
+            cursor.execute("SELECT tablename FROM pg_tables WHERE schemaname not like 'pg_%' and schemaname != 'information_schema'")
+
+            if not [ i for i in cursor.fetchall() if "vlog" in i]:
+                cursor.execute("CREATE TABLE vlog (guild_name varchar(255),guild_id bigint,ch_id bigint,message_id bigint,unix bigint)")
+                self.psql.commit()
     
     @commands.Cog.listener()
     async def on_voice_state_update(self,member:discord.Member,before:discord.VoiceState,after:discord.VoiceState):
         if after.channel!=before.channel:
             if before.channel!=None and len(before.channel.members)==0 and len(before.channel.changed_roles)<=1:
                 sendch=member.guild.system_channel
-                with self.psql:
-                    with self.psql.cursor() as cursor:
-                        sql=f"SELECT unix, message_id FROM {TABLENAME} WHERE guild_id = %s AND ch_id = %s"
-                        cursor.execute(sql,(f"{member.guild.id}",f"{before.channel.id}"))
-                        tmp=cursor.fetchone()
-                retime=int(time.time())-tmp[0]
-                delme=await sendch.fetch_message(tmp[1])
-                await delme.delete()
+                with self.psql.cursor() as cursor:
+                    sql=f"SELECT unix, message_id FROM {TABLENAME} WHERE guild_id = %s AND ch_id = %s"
+                    cursor.execute(sql,(f"{member.guild.id}",f"{before.channel.id}"))
+                    tmp=cursor.fetchone()
+                if tmp is not None:
+                    retime=int(time.time())-tmp[0]
+                    delme=await sendch.fetch_message(tmp[1])
+                    await delme.delete()
+                else:
+                    retime=0
                 data={
                     "title":f"{before.channel}",
                     "fields":[
@@ -70,17 +77,16 @@ class command(commands.Cog):
                     ]
                 }
                 message=await sendch.send(content="@everyone",embed=discord.Embed.from_dict(data=data))
-                with self.psql:
-                    with self.psql.cursor() as cursor:
-                        ssql=f"SELECT unix FROM {TABLENAME} WHERE guild_id = %s AND ch_id = %s"
-                        cursor.execute(ssql,(f"{member.guild.id}",f"{after.channel.id}"))
-                        if cursor.fetchone() is None:
-                            sql=f"INSERT INTO {TABLENAME} (guild_name, guild_id, ch_id, message_id, unix) VALUES (%s, %s, %s, %s, %s)"
-                            cursor.execute(sql,(f"{member.guild.name}",f"{member.guild.id}",f"{after.channel.id}", f"{message.id}", f"{unix}"))
-                        else:
-                            sql=f"UPDATE {TABLENAME} SET message_id = %s, unix = %s WHERE guild_id = %s AND ch_id = %s"
-                            cursor.execute(sql,(f"{message.id}", f"{unix}", f"{member.guild.id}", f"{after.channel.id}"))
-                    self.psql.commit()
+                with self.psql.cursor() as cursor:
+                    ssql=f"SELECT unix FROM {TABLENAME} WHERE guild_id = %s AND ch_id = %s"
+                    cursor.execute(ssql,(f"{member.guild.id}",f"{after.channel.id}"))
+                    if cursor.fetchone() is None:
+                        sql=f"INSERT INTO {TABLENAME} (guild_name, guild_id, ch_id, message_id, unix) VALUES (%s, %s, %s, %s, %s)"
+                        cursor.execute(sql,(f"{member.guild.name}",f"{member.guild.id}",f"{after.channel.id}", f"{message.id}", f"{unix}"))
+                    else:
+                        sql=f"UPDATE {TABLENAME} SET message_id = %s, unix = %s WHERE guild_id = %s AND ch_id = %s"
+                        cursor.execute(sql,(f"{message.id}", f"{unix}", f"{member.guild.id}", f"{after.channel.id}"))
+                self.psql.commit()
 
 async def setup(bot:commands.Bot):
     await bot.add_cog(command(bot))
